analysis routine changed in depth, old analysis routines removed, other files renamed
