+++ grid_shape/analysis.py
@@ -1,45 +1,9 @@
 import matplotlib.pyplot as plt
 import numpy as np
 import os
-import tale
+import diff_spec
+import utils_analysis as ua
 
-class Event:
-    def __init__(self,f1,f2, step, name):
-        p2px, p2py, p2pz, p2ptot = np.loadtxt(f1)
-        self.p2px = p2px
-        self.p2py = p2py
-        self.p2pz = p2pz
-        self.p2ptot = p2ptot
-        # JobName,Primary,Energy,Zenith,Azimuth,XmaxDistance,SlantXmax,RandomCore[0],RandomCore[1],RandomAzimuth,HadronicModel
-        A = open(f2).readlines()[0]
-        A = A.strip().split()
-        self.jobname = A[0]
-        self.primary = A[1]
-        self.energy = np.float32(A[2])
-        self.zenith = np.float32(A[3])
-        self.azimuth = np.float32(A[4])
-        self.xmax_distance = np.float32(A[5])
-        self.slant_xmax = np.float32(A[6])
-        self.random_core0 = np.float32(A[7])
-        self.random_core1 = np.float32(A[8])
-        self.random_azimuth = np.float32(A[9])
-        self.hadronic_model = A[10]
-        self.step = np.float32(step)
-        self.name = name
-        self.init_layout()
-
-    def init_layout(self):
-        if "hexhex" in self.name:
-            self.layout = "hexhex"
-        elif "rect" in self.name:
-            self.layout = "rect"
-        elif "trihex" in self.name:
-            self.layout = "trihex"
-        else:
-            self.layout = "unknown"
-
-    def is_triggered1(self, threshold):
-        return self.p2ptot > threshold
 
 
 # read files
@@ -47,130 +11,96 @@ class Event:
 #path = "/Users/kotera/BROQUE/Data_GRAND/Matias/InterpolationOutputExample/"
 #path = "/Users/kotera/BROQUE/Data_GRAND/Matias/StshpLibrary02/"
 #path = "/Users/kotera/BROQUE/Data_GRAND/Matias/P2PdataNew/"
-path = "/Users/kotera/BROQUE/Data_GRAND/Matias/StshpLibaryHDF5-Grids/"
-#path = "/Users/benoitl/Documents/GRAND/P2PDataNew/P2PdataNew/"
+#path = "/Users/kotera/BROQUE/Data_GRAND/Matias/StshpLibaryHDF5-Grids/"
+path = "/Users/benoitl/Documents/GRAND/StshpLibaryHDF5-Grids/"
 
-plot_path = '/Users/kotera/BROQUE/Plots_GRAND/'
-#plot_path = '/Users/benoitl/Documents/GRAND/P2PDataNew/plots'
+#plot_path = '/Users/kotera/BROQUE/Plots_GRAND/'
+plot_path = '/Users/benoitl/Documents/GRAND/plots_tests'
+events_data_dir = "/Users/benoitl/Documents/GRAND/event_data"
+
 os.makedirs(plot_path, exist_ok=True)
+os.makedirs(events_data_dir, exist_ok=True)
 
-
-ev_list = []
-count = 0
-
-#for subdir in os.listdir(path)[0:25000]:
-for subdir in os.listdir(path)[0:5000]:
-    if os.path.isdir(os.path.join(path, subdir)):
-        list_fn = os.listdir(os.path.join(path, subdir))        
-        for fn in list_fn:
-            if fn[-6:] == "P2Pdat":
-                f1 = os.path.join(path, subdir, fn)
-                f2 = os.path.join(path, subdir, fn+'.showerinfo')
-                step  = fn.split("_")[-2]
-
-                ev_list.append(Event(f1, f2, step, fn))
-
-    count += 1 
-    if(count % 100 == 0):
-        print("Event #{} done".format(count))
-
-# select triggered antennas and events
 threshold = 30
-# is_triggered_list = [sum(ev.is_triggered(75)) for ev in ev_list if "voltage" in ev.name]
 
-for ev in ev_list:
-    if "voltage" in ev.name:
-        ev.num_triggered = sum(ev.is_triggered1(threshold))
-        ev.is_triggered2 = (ev.num_triggered > 10)
+ev_select_rect_file = os.path.join(events_data_dir, 'ev_select_rect.npy')
+ev_select_hexhex_file = os.path.join(events_data_dir, 'ev_select_hexhex.npy')
+ev_select_trihex_file = os.path.join(events_data_dir, 'ev_select_trihex.npy')
 
-# A = [(ev.num_triggered, ev.energy, ev.step, ev.primary, ev.layout, ev.zenith) for  ev in ev_list if "voltage" in ev.name]
+## Check that A_rect exits
 
-# make array with each event information 
-A = [(ev.num_triggered, ev.energy, ev.step, ev.layout, ev.zenith) for  ev in ev_list if "voltage" in ev.name and ev.primary == "Proton"]
+do_make_ev_list = (
+    ua.isnot_ev_select(ev_select_rect_file) *
+    ua.isnot_ev_select(ev_select_hexhex_file) *
+    ua.isnot_ev_select(ev_select_trihex_file)
+)
 
-A_rect = [
-    (ev.num_triggered, ev.energy, ev.step, ev.zenith, ev.is_triggered2) for  ev in ev_list
-    if "voltage" in ev.name
-    and ev.primary == "Proton"
-    and ev.layout == 'trihex'
-]
-A_hexhex = [
-    (ev.num_triggered, ev.energy, ev.step, ev.zenith, ev.is_triggered2) for  ev in ev_list
-    if "voltage" in ev.name
-    and ev.primary == "Proton"
-    and ev.layout == 'hexhex'
-]
+if do_make_ev_list:
+    ev_list = ua.make_ev_list(path)
+    # select triggered antennas and events
+    
+    for ev in ev_list:
+        if "voltage" in ev.name:
+            ev.num_triggered = sum(ev.is_triggered1(threshold))
+            ev.is_triggered2 = (ev.num_triggered > 10)
 
-A_rect = np.array(A_rect)
-A_hexhex = np.array(A_hexhex)
+
+    ev_select_rect = ua.make_ev_select(
+        ev_list,
+        'rect',
+        'Proton',
+        ev_select_rect_file
+    )
+
+    ev_select_hexhex = ua.make_ev_select(
+        ev_list,
+        'hexhex',
+        'Proton',
+        ev_select_hexhex_file
+    )
+    ev_select_trihex = ua.make_ev_select(
+        ev_list,
+        'trihex',
+        'Proton',
+        ev_select_trihex_file
+    )
+
+else:
+    ev_select_rect = np.load(ev_select_rect_file)
+    ev_select_hexhex = np.load(ev_select_hexhex_file)
+    ev_select_trihex = np.load(ev_select_trihex_file)
+
+
+
 
 # calculate mean and variance of triggered antenna numbers in each zenith angle and energy bins 
-enerbins = np.unique(A_rect[:,1])
+enerbins = np.unique(ev_select_rect[:,1])
 #zenbins = 180-np.unique(A_rect[:,3])
 zenbins = np.array([94.77,95.74,97.18,98.21,99.59,101.54, 104.48, 106.6, 109.47, 113.58,120,132])
 zenbins = 180. - zenbins
 #zenbins = [94,100,105,110,120,131]
-stepbins = np.unique(A_rect[:,2])
+stepbins = np.unique(ev_select_rect[:,2])
 
-meanNtrig_ener = []
-varNtrig_ener = []
 
-for istep, step in enumerate(stepbins):
-    meanNtrig_step = []
-    varNtrig_step = []
 
-    for iener, ener in enumerate(enerbins):
-        meanNtrig_zen = []
-        varNtrig_zen = []
-    
-        #for izen in range(0, len(zenbins)-1):
-        for izen, zen in enumerate(zenbins):
-            ind = np.where((A_rect[:,1] == ener) * (A_rect[:,2] == step)
-                * (np.abs(A_rect[:,3]-(180-zen)) < 0.5) * (A_rect[:,0] > 0))
-                #* (A_rect[:,3] == 180-zen)* (A_rect[:,0] > 0))
-                #* (A_rect[:,3] >= zenbins[izen]) * (A_rect[:,3] < zenbins[izen+1]))
-            #print(ind)
-            if (len(ind[0]) == 0):
-                meanNtrig_zen.append(0)
-                varNtrig_zen.append(0)   
-            else:
-                meanNtrig_zen.append(np.mean(A_rect[ind[0],0]))
-                varNtrig_zen.append(np.var(A_rect[ind[0],0]))
+meanNtrig_ener, varNtrig_ener = ua.compute_meanNtrig(
+    stepbins,
+    enerbins,
+    zenbins,
+    ev_select_rect
+)
 
-        meanNtrig_step.append(meanNtrig_zen)
-        varNtrig_step.append(varNtrig_zen)
 
-    meanNtrig_ener.append(meanNtrig_step)
-    varNtrig_ener.append(varNtrig_step)
-
-meanNtrig_ener = np.array(meanNtrig_ener)
-varNtrig_ener = np.array(varNtrig_ener)
 
 
 # calculate mean and variance of triggered event numbers in each zenith angle and energy bins 
 
-Ntrig2_ener = []
-
-for istep, step in enumerate(stepbins):
-    Ntrig2_step = []
-
-    for iener, ener in enumerate(enerbins):
-        Ntrig2_zen = []
-    
-        #for izen in range(0, len(zenbins)-1):
-        for izen, zen in enumerate(zenbins):
-            ind = np.where((A_rect[:,1] == ener) * (A_rect[:,2] == step) 
-                *(np.abs(A_rect[:,3]-(180-zen)) < 0.5))
-                #* (A_rect[:,3] == 180-zen))
-                #* (A_rect[:,3] >= zenbins[izen]) * (A_rect[:,3] < zenbins[izen+1]))
-            Ntrig2_zen.append(sum(A_rect[ind[0],4])/np.size(A_rect[ind[0],4]))
-
-        Ntrig2_step.append(Ntrig2_zen)
-
-    Ntrig2_ener.append(Ntrig2_step)
-
-Ntrig2_ener = np.array(Ntrig2_ener)
-
+Ntrig2_ener = ua.compute_trig_rate(
+    stepbins,
+    enerbins,
+    zenbins,
+    ev_select_rect
+)
 
 
 # plot Ntriggered antennas vs energies for fixed steps
@@ -204,7 +134,7 @@ for izen in range(0, len(zenbins)-1):
     plt.figure(izen+4) 
     plt.clf()
     for istep, step in enumerate(stepbins):
-        plt.errorbar(enerbins, meanNtrig_ener[istep,:,izen], yerr=sqrt(varNtrig_ener[istep,:,izen]), 
+        plt.errorbar(enerbins, meanNtrig_ener[istep,:,izen], yerr=np.sqrt(varNtrig_ener[istep,:,izen]), 
             fmt=sym_list[istep], capsize=2, alpha=0.7, label='step = %d m'%(np.int32(step)))
         #plt.errorbar(enerbins, Ntrig2_ener[istep,:,izen],  
          #   fmt=sym_list[istep], capsize=2, alpha=0.7)
@@ -220,7 +150,7 @@ for iener, ener in enumerate(enerbins):
     plt.figure(iener) 
     plt.clf()
     for istep, step in enumerate(stepbins):
-        plt.errorbar(zenbins, meanNtrig_ener[istep,iener,:], yerr=sqrt(varNtrig_ener[istep,iener,:]), 
+        plt.errorbar(zenbins, meanNtrig_ener[istep,iener,:], yerr=np.sqrt(varNtrig_ener[istep,iener,:]), 
             fmt=sym_list[istep], capsize=2, alpha=0.7, label='step = %d m'%(np.int32(step)))
         #plt.errorbar(enerbins, Ntrig2_ener[istep,:,izen], 
         #    fmt=sym_list[izen], capsize=2)
@@ -324,7 +254,7 @@ for iener, ener in enumerate(enerbins):
         for izen, zen in enumerate(zenbins):
             rate[istep, iener, izen] = (
                 Ntrig2_ener[istep,iener,izen] * 
-                tale.tale_diff_flux(ener*1e18) * 
+                diff_spec.tale_diff_flux(ener*1e18) * 
                 delta_E[iener] *1e18 * delta_omega[izen] *
                 area[istep]
             )
