+++ grid_shape/diff_spec.py
